pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langcode"
version = "0.1.0"
description = "A Claude Code/OpenCode clone built with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "tree-sitter>=0.20.0",
]

[project.scripts]
langcode = "langcode.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
